+++ src/main/java/insane96mcp/iguanatweaksreborn/module/mobs/villager/villagers/Villagers.java
@@ -1,7 +1,9 @@
-package insane96mcp.iguanatweaksreborn.module.mobs;
+package insane96mcp.iguanatweaksreborn.module.mobs.villager.villagers;
 
 import insane96mcp.iguanatweaksreborn.InsaneSurvivalOverhaul;
 import insane96mcp.iguanatweaksreborn.module.Modules;
+import insane96mcp.iguanatweaksreborn.module.farming.crops.Crops;
+import insane96mcp.iguanatweaksreborn.module.mobs.villager.SerializableTrade;
 import insane96mcp.insanelib.base.Feature;
 import insane96mcp.insanelib.base.Label;
 import insane96mcp.insanelib.base.LoadFeature;
@@ -14,19 +16,51 @@ import net.minecraft.world.entity.ai.gossip.GossipType;
 import net.minecraft.world.entity.npc.Villager;
 import net.minecraft.world.entity.npc.VillagerProfession;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.Items;
 import net.minecraftforge.event.entity.living.LivingEvent;
 import net.minecraftforge.event.entity.living.MobEffectEvent;
 import net.minecraftforge.eventbus.api.Event;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Map;
 import java.util.UUID;
+import java.util.function.Supplier;
 
-@Label(name = "Villagers Nerfs", description = "Small changes to villagers to make them less OP")
+@Label(name = "Villagers", description = "Nerfs to villagers + change their trades via json config")
 @LoadFeature(module = Modules.Ids.MOBS)
-public class VillagersNerfs extends Feature {
+public class Villagers extends Feature {
 
 	private static final String CURE_DISCOUNT_REMOVED = InsaneSurvivalOverhaul.RESOURCE_PREFIX + "cure_discount_removed";
+
+	public static final Supplier<ArrayList<SerializableTrade>> VILLAGER_TRADES_DEFAULT = () -> new ArrayList<>(List.of(
+			new SerializableTrade(new ItemStack(Items.EMERALD, 2), new ItemStack(Items.WHEAT_SEEDS), 4),
+			new SerializableTrade(new ItemStack(Items.EMERALD, 2), new ItemStack(Crops.CARROT_SEEDS.get()), 4),
+			new SerializableTrade(new ItemStack(Items.EMERALD, 2), new ItemStack(Crops.ROOTED_POTATO.get()), 4),
+			new SerializableTrade(new ItemStack(Items.EMERALD, 2), new ItemStack(Items.BEETROOT_SEEDS), 4),
+			new SerializableTrade(new ItemStack(Items.EMERALD, 1), new ItemStack(Items.BROWN_MUSHROOM), 4),
+			new SerializableTrade(new ItemStack(Items.EMERALD, 1), new ItemStack(Items.RED_MUSHROOM), 4),
+			new SerializableTrade(new ItemStack(Items.EMERALD, 3), new ItemStack(Items.MELON), 2),
+			new SerializableTrade(new ItemStack(Items.EMERALD, 3), new ItemStack(Items.PUMPKIN), 2),
+			new SerializableTrade(new ItemStack(Items.EMERALD, 1), new ItemStack(Items.VINE), 4),
+			new SerializableTrade(new ItemStack(Items.EMERALD, 1), new ItemStack(Items.LILY_PAD, 2), 4),
+			new SerializableTrade(new ItemStack(Items.EMERALD, 2), new ItemStack(Items.POINTED_DRIPSTONE, 2), 3),
+			new SerializableTrade(new ItemStack(Items.EMERALD, 2), new ItemStack(Items.SEA_PICKLE), 3),
+			new SerializableTrade(new ItemStack(Items.EMERALD, 3), new ItemStack(Items.CACTUS), 3),
+			new SerializableTrade(new ItemStack(Items.EMERALD, 3), new ItemStack(Items.KELP), 3),
+			new SerializableTrade(new ItemStack(Items.EMERALD, 5), new ItemStack(Items.ACACIA_SAPLING), 3),
+			new SerializableTrade(new ItemStack(Items.EMERALD, 5), new ItemStack(Items.BIRCH_SAPLING), 3),
+			new SerializableTrade(new ItemStack(Items.EMERALD, 5), new ItemStack(Items.CHERRY_SAPLING), 3),
+			new SerializableTrade(new ItemStack(Items.EMERALD, 5), new ItemStack(Items.SPRUCE_SAPLING), 3),
+			new SerializableTrade(new ItemStack(Items.EMERALD, 5), new ItemStack(Items.OAK_SAPLING), 3),
+			new SerializableTrade(new ItemStack(Items.EMERALD, 5), new ItemStack(Items.DARK_OAK_SAPLING), 3),
+			new SerializableTrade(new ItemStack(Items.EMERALD, 5), new ItemStack(Items.JUNGLE_SAPLING), 3),
+			new SerializableTrade(new ItemStack(Items.EMERALD, 5), new ItemStack(Items.NAUTILUS_SHELL), 5),
+			new SerializableTrade(new ItemStack(Items.EMERALD, 3), new ItemStack(Items.TROPICAL_FISH_BUCKET), 4),
+			new SerializableTrade(new ItemStack(Items.EMERALD, 3), new ItemStack(Items.PUFFERFISH_BUCKET), 4)
+	));
+	public static final ArrayList<VillagerTrade> trades = new ArrayList<>();
 
 	@Config
 	@Label(name = "Lock Trades", description = "If true, villagers will be given 1 trading experience as soon as they choose their job to lock the trades.")
@@ -50,7 +84,7 @@ public class VillagersNerfs extends Feature {
 	@Label(name = "Remove Bad Omen", description = "If true, the effect can no longer be applied to entities")
 	public static Boolean removeBadOmen = false;
 
-	public VillagersNerfs(Module module, boolean enabledByDefault, boolean canBeDisabled) {
+	public Villagers(Module module, boolean enabledByDefault, boolean canBeDisabled) {
 		super(module, enabledByDefault, canBeDisabled);
 	}
 
@@ -96,7 +130,7 @@ public class VillagersNerfs extends Feature {
 	}
 
 	public static int clampSpecialPrice(int specialPriceDiff, final ItemStack baseCostA) {
-		if (!isEnabled(VillagersNerfs.class)
+		if (!isEnabled(Villagers.class)
 				|| maxDiscount == 1d)
 			return specialPriceDiff;
 
@@ -107,7 +141,7 @@ public class VillagersNerfs extends Feature {
 	}
 
 	public static int clampDemand(int demand, int maxUses) {
-		if (!isEnabled(VillagersNerfs.class)
+		if (!isEnabled(Villagers.class)
 				|| !clampNegativeDemand)
 			return demand;
 
@@ -115,7 +149,7 @@ public class VillagersNerfs extends Feature {
 	}
 
 	public static void lockTrades(Villager villager) {
-		if (!isEnabled(VillagersNerfs.class)
+		if (!isEnabled(Villagers.class)
 				|| !lockTrades)
 			return;
 
@@ -124,6 +158,6 @@ public class VillagersNerfs extends Feature {
 	}
 
 	public static boolean shouldConvertVillagerToZombie() {
-		return isEnabled(VillagersNerfs.class) && alwaysConvertZombie;
+		return isEnabled(Villagers.class) && alwaysConvertZombie;
 	}
 }